+++ resource/Changelog.md
@@ -2,6 +2,12 @@
 
 ## v1.1.4
 
+**feat**
+
+- [ ] support register and run plugins
+
+**update**
+
 - [x] update readme
-- [x] support set env on run git commands
+- [x] support set env on run some github commands
 - [x] `php serve` command support idea `http-client.env.json`